+++ checksumdialog.h
@@ -39,6 +39,13 @@ public:
 
     /// Whether the user created new files during the dialog execution.
     bool hasCreatedNewFiles() const;
+
+    /// Sets the hash algorithm which is pre-selected for the dialog.
+    ///
+    /// Note: MD4 and Keccak variants are currently not supported.
+    void setPreSelectedAlgorithm(const QCryptographicHash::Algorithm algorithm);
+signals:
+    void hashAlgorithmChanged(const QCryptographicHash::Algorithm algorithm);
 private slots:
     void rbAlgorithmToggled(bool checked);
     void btnCalculateClicked();