+++ changelog.md
@@ -19,6 +19,10 @@ improvements may be omitted.)_
   BLAKE2b-256, BLAKE2b-384, BLAKE2b-512, BLAKE2s-128, BLAKE2s-160, BLAKE2s-224,
   and BLAKE2s-256).
 
+* __[improvement]__
+  The checksum dialog will now remember the last used hash algorithm for the
+  current session and pre-select that whenever the checksum dialog is opened.
+
 ## Version 0.1.5 (2026-05-04)
 
 * __[fix]__